[#10] Simplify store interface and update finder
When processing the segment, we need only one busStop's route info, the first one, because the former UNAUTO service only needs the order and the route to retrieve the time-to-arrive. In the former service, those order and route will be in the header of the response, but we are merging all together.